Refactored to server actions
